LayoutTests/accessibility/ios-simulator/live-regions/live-region-removal.html

<!DOCTYPE HTML><!-- webkit-test-runner [ IsAriaLiveRegionManagementEnabled=true ] -->
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<script src="../../../resources/accessibility-helper.js"></script>
<meta http-equiv="content-language" content="en">
</head>
<body>

<div id="polite-region" aria-live="polite" aria-relevant="removals">
    Hello 
    <p id="text">world</p>
</div>

<script>
var output = "This test verifies that live region notification contains the removal attribute with aria-relevant='removals'.\n\n";
var testFinished = false;

function notificationCallback(object, notification, userInfo) {
    if (notification == "AXAnnouncementRequested") {
        output += `Announcement received, notification: ${notification}, attributed string: ${userInfo["AXAnnouncementRequested"]}\n`;

        testFinished = true;
    }
}

if (accessibilityController) {
    window.jsTestIsAsync = true;
    accessibilityController.addNotificationListener(notificationCallback);
    accessibilityController.accessibleElementById("polite-region");

    document.getElementById('text').remove();
 
    setTimeout(async function() {
        await waitFor(() => testFinished);
        accessibilityController.removeNotificationListener();
        debug(output);
        finishJSTest();
    }, 0);
}
</script>
</body>
</html>
