LayoutTests/accessibility/ios-simulator/live-regions/live-region-removal-expected.txt

This test verifies that live region notification contains the removal attribute with aria-relevant='removals'.

Announcement received, notification: AXAnnouncementRequested, attributed string: world{
    UIAccessibilitySpeechAttributeAnnouncementPriority = UIAccessibilityPriorityLow;
    UIAccessibilitySpeechAttributeLanguage = en;
    UIAccessibilityTokenIsLiveRegionRemoval = 1;
    UIAccessibilityTokenLiveRegionAnnouncement = 1;
}

PASS successfullyParsed is true

TEST COMPLETE
Hello
